matcher.py: Skip elements that only match themselves
Previously, the XML output included elements that
only match themselves to indicate that they were
checked.
This makes handling the output more complicated,
as a user is generally only interested in getting
the elements of the repository that actually
match something. We therefore adapt the output
generation to only include matches that match
something other than themselves.